Move bodhi-messages in its own directory
