+++ bodhi-messages/README.rst
@@ -1,0 +1,4 @@
+Bodhi Messages
+==============
+
+This package contains the schema for messages published by Bodhi.
