+++ bodhi-messages/setup.cfg
@@ -1,0 +1,55 @@
+[metadata]
+name = bodhi-messages
+version = attr: bodhi.messages.__version__
+description = JSON schema for messages sent by Bodhi
+long_description = file: README.rst
+maintainer = Fedora Infrastructure Team
+maintainer_email = infrastructure@lists.fedoraproject.org
+platforms = Fedora, GNU/Linux
+url = https://github.com/fedora-infra/bodhi
+keywords = fedora, fedora-messaging
+license = GPLv2+
+# Possible options are at https://pypi.python.org/pypi?%3Aaction=list_classifiers
+classifiers =
+    Development Status :: 5 - Production/Stable
+    Intended Audience :: Developers
+    Intended Audience :: System Administrators
+    License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)
+    Operating System :: POSIX :: Linux
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Topic :: System :: Software Distribution
+
+[options]
+zip_safe = False
+include_package_data = True
+packages = find_namespace:
+namespace_packages = bodhi
+install_requires =
+    fedora_messaging
+
+[options.entry_points]
+fedora.messages =
+    bodhi.buildroot_override.tag.v1 = bodhi.messages.schemas.buildroot_override:BuildrootOverrideTagV1
+    bodhi.buildroot_override.untag.v1 = bodhi.messages.schemas.buildroot_override:BuildrootOverrideUntagV1
+    bodhi.errata.publish.v1 = bodhi.messages.schemas.errata:ErrataPublishV1
+    bodhi.compose.complete.v1 = bodhi.messages.schemas.compose:ComposeCompleteV1
+    bodhi.compose.composing.v1 = bodhi.messages.schemas.compose:ComposeComposingV1
+    bodhi.compose.start.v1 = bodhi.messages.schemas.compose:ComposeStartV1
+    bodhi.compose.sync.done.v1 = bodhi.messages.schemas.compose:ComposeSyncDoneV1
+    bodhi.compose.sync.wait.v1 = bodhi.messages.schemas.compose:ComposeSyncWaitV1
+    bodhi.repo.done.v1 = bodhi.messages.schemas.compose:RepoDoneV1
+    bodhi.update.comment.v1 = bodhi.messages.schemas.update:UpdateCommentV1
+    bodhi.update.complete.stable.v1 = bodhi.messages.schemas.update:UpdateCompleteStableV1
+    bodhi.update.complete.testing.v1 = bodhi.messages.schemas.update:UpdateCompleteTestingV1
+    bodhi.update.status.testing.v1 = bodhi.messages.schemas.update:UpdateReadyForTestingV1
+    bodhi.update.edit.v1 = bodhi.messages.schemas.update:UpdateEditV1
+    bodhi.update.eject.v1 = bodhi.messages.schemas.update:UpdateEjectV1
+    bodhi.update.karma.threshold.v1 = bodhi.messages.schemas.update:UpdateKarmaThresholdV1
+    bodhi.update.request.revoke.v1 = bodhi.messages.schemas.update:UpdateRequestRevokeV1
+    bodhi.update.request.stable.v1 = bodhi.messages.schemas.update:UpdateRequestStableV1
+    bodhi.update.request.testing.v1 = bodhi.messages.schemas.update:UpdateRequestTestingV1
+    bodhi.update.request.unpush.v1 = bodhi.messages.schemas.update:UpdateRequestUnpushV1
+    bodhi.update.request.obsolete.v1 = bodhi.messages.schemas.update:UpdateRequestObsoleteV1
+    bodhi.update.requirements_met.stable.v1 = bodhi.messages.schemas.update:UpdateRequirementsMetStableV1
