+++ bodhi-messages/tests/schemas/test_compose.py
@@ -24,7 +24,7 @@ from bodhi.messages.schemas.compose import (
     ComposeSyncDoneV1,
     RepoDoneV1,
 )
-from bodhi.tests.messages.utils import check_message
+from ..utils import check_message
 
 
 class TestComposeMessage:
